app/api/requirements.txt
fastapi==0.89.1
pillow==9.5.0
slowapi==0.1.7
fastapi-pagination==0.11.2
python-multipart==0.0.5
sqlalchemy==1.4.46
sqlalchemy-utils==0.39.0
pydantic==1.10.4
psycopg2==2.9.5
python-jose==3.3.0
cryptography==39.0.0
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
uvicorn[standard]==0.20.0
jinja2==3.1.2
